+++ resources/views/components/emails/footer-html.blade.php
@@ -15,6 +15,9 @@
     @lang('ig-common::layouts.email.generated-at', ['url' => $url])<br />
     @lang('ig-common::layouts.email.generator', ['generator' => InternetGuru\LaravelCommon\Support\Helpers::getAppInfo()])<br />
     @lang('ig-common::layouts.email.requested-from', ['ip' => $ip, 'timezone' => $timezone ?? 'n/a'])
+    @if ($noreplyMessage)
+        <br />{{ $noreplyMessage }}
+    @endif
 </p>
 <p>
     @lang('ig-common::layouts.provider', ['link' => $providerLink, 'year' => date('Y')])